# docs/api/openapi.json
{"openapi":"3.1.0","info":{"title":"Anclora RAG API","description":"API REST bilingüe para consultar y administrar el sistema Anclora RAG.\n\nThis bilingual REST API allows you to query and manage the Anclora RAG system.","version":"1.0.0"},"paths":{"/health":{"get":{"summary":"Health check / Verificación de estado","description":"Verifica el estado general de la API, reporta la versión y el estado de los servicios críticos.\n\nChecks the overall API health, the deployed version, and reports the status of critical services.","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HealthResponse"}}}}}}},"/chat":{"post":{"summary":"Conversación con el RAG / Chat with RAG","description":"Envía una consulta en español o inglés al motor RAG y recibe una respuesta contextualizada.\n\nSend a Spanish or English query to the RAG engine and receive a contextualised answer.","operationId":"chat_with_rag_chat_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/upload":{"post":{"summary":"Subir documento / Upload document","description":"Permite cargar un documento válido para que se procese e indexe en la base de conocimiento del RAG.\n\nUpload a valid document so it can be processed and indexed into the RAG knowledge base.","operationId":"upload_document_upload_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_document_upload_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/documents":{"get":{"summary":"Listar documentos / List documents","description":"Devuelve la lista de fuentes únicas presentes en la base vectorial del RAG.\n\nReturns the list of unique sources available in the RAG vector database.","operationId":"list_documents_documents_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"HTTPBearer":[]}]}},"/documents/{filename}":{"delete":{"summary":"Eliminar documento / Delete document","description":"Elimina la referencia del documento indicado de la base de conocimiento RAG.\n\nDeletes the specified document reference from the RAG knowledge base.","operationId":"delete_document_documents__filename__delete","security":[{"HTTPBearer":[]}],"parameters":[{"name":"filename","in":"path","required":true,"schema":{"type":"string","title":"Filename"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"Body_upload_document_upload_post":{"properties":{"file":{"type":"string","format":"binary","title":"File"}},"type":"object","required":["file"],"title":"Body_upload_document_upload_post"},"ChatRequest":{"properties":{"message":{"type":"string","title":"Message","description":"Mensaje de la consulta o pregunta que se enviará al motor RAG.\nQuery message or question that will be processed by the RAG engine.","example":"¿Cuál es la política de respaldo de datos?"},"max_length":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Max Length","description":"Longitud máxima permitida para el mensaje. Ajusta el límite si necesitas respuestas extensas.\nMaximum number of characters allowed in the message. Increase the limit for longer prompts.","default":1000,"example":800},"language":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Language","description":"Idioma preferido para la respuesta (`es` o `en`).\nPreferred response language (`es` or `en`).","default":"es","example":"en"}},"type":"object","required":["message"],"title":"ChatRequest","example":{"language":"en","max_length":600,"message":"Provide a summary of the onboarding process."}},"ChatResponse":{"properties":{"response":{"type":"string","title":"Response","description":"Respuesta generada por el sistema RAG. Puede incluir texto enriquecido.\nAnswer generated by the RAG system. The text may contain references or summaries.","example":"La política de respaldo realiza copias incrementales cada 24 horas..."},"status":{"type":"string","title":"Status","description":"Estado de la operación (`success` o `error`).\nOperation status (`success` or `error`).","example":"success"},"timestamp":{"type":"string","title":"Timestamp","description":"Marca temporal ISO 8601 de la respuesta.\nISO 8601 timestamp of the response.","example":"2024-05-12T14:32:10.456789"}},"type":"object","required":["response","status","timestamp"],"title":"ChatResponse","example":{"response":"Las copias de seguridad se ejecutan automáticamente cada noche...","status":"success","timestamp":"2024-05-12T14:32:10.456789"}},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"HealthResponse":{"properties":{"status":{"type":"string","title":"Status","description":"Estado global del sistema.\nOverall health status of the system.","example":"healthy"},"version":{"type":"string","title":"Version","description":"Versión desplegada de la API.\nDeployed API version.","example":"1.0.0"},"services":{"type":"object","title":"Services","description":"Mapa de servicios críticos y su estado operativo.\nMapping of critical services and their operational status.","example":{"chroma_db":"healthy","ollama":"healthy"}}},"type":"object","required":["status","version","services"],"title":"HealthResponse","example":{"services":{"chroma_db":"healthy","ollama":"healthy","rag_system":"healthy"},"status":"healthy","version":"1.0.0"}},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"HTTPBearer":{"type":"http","scheme":"bearer"}}}}